vsr: repair headers&prepares even when lagging behind the primary
Our repair procedure have a condition where we don't initiate repair
until our head op is at least commit_max. This feels suspicious to me,
as it feels like this _could_ live-lock a replica, if the commit max
advances faster than the op (via commit messages).

I _think_ this can be safely relaxed to requiring that the current op is
from the current view --- if we have _some_ op from a view, that gives
us a strong cryptographic proof tha everything that hash-chains to it
is, in fact, a part of the view in question.