vsr: tone down false warning in clock
m0 (time of ping) can legitimately be prior to window start if the reset
happened right between a ping and a pong, which is by far the likelier
explanation than a restart.

On the other hand, m2 has to be within the window, as this compares two
reads of the monotonic clock of the same process.