bus: adaptively batch messages in the send queue
Before this PR, sending a message meant adding it as is to a message
queue. Given that the queue is only 4 messages long, this could easily
overflow the queue if the messages are small. In particular, if a
replica accumulated a lot of bytes in its tcp receive buffer, then it'll
receive a bunch of messages almost simultaneously, and can overflow the
send queue without givin a chance for the kernel to even copy the bytes.

To combat this problem, introduce batching of messages, in an approach
dual to that on the receive path:

- if the queue is empty, send a message even if it is tiny
- if there's already a queued message, try to batch with it
    - batching involves an extra memcpy, but this is fine, as we can't
      send the bytes over at the moment anyway