vsr: replicate committed messages
Previously, only messages advancing op and not known to be committed
were replicated. This played badly with a pipeline: if two prepares get
reordered on _one_ replica, it creates a bubble that rides until the end
of the ring, triggering repairs on any replicas along the way.

With this commit, we replicate any prepare that we haven't seen before.

I originally considered explicitly amounting for a pipeline, and
repairing after `op -| pipeline`, but I think that would be worse.

Consider a scenario where half of the cluster gets odd prepares, and
half of the cluster gets even ones. Here, if a replica from the first
half gets a lucky even prepare, it can be well-advised to forward it
along.

Another case: consider replica re-joining the cluster after a delay.
This replica will be requesting all prepares, and it, at the first
sight, seems wasteful to replicate them all. But actually, as the
replica were missing, it is likely that anyone in front of it is now
missing these prepares either! So it actually is good if it can replay
them.