bus:  batch small messages when sending over socket
Before this PR, sending a message meant adding it as is to a message
queue. Given that the queue is only 4 messages long, this could easily
overflow the queue if the messages are small. In particular, if a
replica accumulated a lot of bytes in its tcp receive buffer, then it'll
receive a bunch of messages almost simultaneously, and can overflow the
send queue without givin a chance for the kernel to even copy the bytes.

To combat this problem, try submitting messages to the kernel as soon as
possible. That is, whenever we want to send a message, do a
synchroneous, non-blocking write to the socket. Most likely, this'll
just copy bytes over to the kernel. If it returns WouldBlock, then fall
back to asynchronous send via io_uring.