vsr: optimize the order of repair of prepares
Instead repairing strictly chronologically, start with commit_min+1, so
that we can commit further.

This is helpful in the following scenario:

- the cluster is under high load,
- in particular, the volume of prepares a backup gets through the
  network overwhelms its disk,
- so a backup doesn't write _all_ incoming prepares to disk
- but still commits them, as they are added to the pipeline_cache

In other words, a replica can end up in a situation where it misses a
lot of prepares prior to its own commit min! Before this commit, what
would happen is that the replica would repair all those prepares before
anything that it actually needs to commit further.

Now, we always try to repair `commit_min+1` first.