vsr: speed up commit during repair
Previously, the on_repair kicked off commit process asynchronously, in
the journal_write callback (through `.repair` call).

This is unnecessary: even before the prepare is written to storage, the
replica keeps it in memory in pipeline cache, and we can commit from
memory.

This potentially is quite harmful --- if a lagging replica catches up
with a cluster and repairs a lot of prepares, it could be the case that,
by the time the write finishes, the prepare is already evicted from the
pipeline cache! This in turn can lead to a situation where up-to-date
replicas commit mostly from memory, while a lagging replica commits by
reading from storage, preventing it from catching up forever.