vsr: disable costly cache map verification
At the moment, we ship production releases with `verify=true`, and I
think this might be a bit too costly for that: effectively, here we are
re-doing a bar-worth of work. While it is ok for constants.verify in
terms of amortized big-O, it might be not ok for latencies.