Journal: Don't write invalid redundant headers during repair
# Overview

To hit this bug:

1. Recover from a WAL with a faulty entry, then
2. repair that entry,
3. then crash and recover again before the WAL wraps and overwrites that entry.

If the bug is hit, the just-repaired slot is recovered as faulty.

# Bug

Suppose we are repairing a faulty prepare:
1. (The dirty/faulty bits are already set.)
2. We start by writing the prepare...
3. Then when the prepare is written, we write the redundant header sector...
4. Then when the redundant header sector is written, we clear the dirty/faulty bits.

The problem is that during step 3, since the faulty bit is set, the sector of redundant headers that we write contains garbage (i.e. "deliberately" invalid headers -- incorrectly so, in this case). This is the code (from `Journal.header_sector()`) which generates the redundant headers to write:

    var i: usize = 0;
    while (i < headers_per_sector) : (i += 1) {
        const slot = Slot{ .index = sector_slot.index + i };

        if (journal.faulty.bit(slot)) {
            // Redundant faulty headers are deliberately written as invalid.
            // This ensures that faulty headers are still faulty when they are read back
            // from disk during recovery. This prevents faulty entries from changing to
            // reserved (and clean) after a crash and restart (e.g. accidentally converting
            // a case `@D` to a `@I` after a restart).
            sector_headers[i] = Header.Prepare.reserved(replica.cluster, i);
            sector_headers[i].checksum = 0; // Invalidate the checksum.
            assert(!sector_headers[i].valid_checksum());
        } else {
            // Write headers from `headers_redundant` instead of `headers` — we need to
            // avoid writing (leaking) a redundant header before its corresponding prepare
            // is on disk.
            sector_headers[i] = journal.headers_redundant[slot.index];
        }
    }

This garbage means that after step 4, if we crash/recover, we can end up with the same slot as faulty, even though we just repaired it.

#### Why didn't the VOPR find this bug?

(Note that with the `JournalChecker`, the VOPR _does_ catch this bug now.)

Redundant headers only influence the test during WAL recovery.

The fault atlas mostly ensures that we don't inject a pattern of WAL faults that the cluster cannot repair.
The caveat of that "mostly" is that the fault atlas doesn't try to prevent `status=recovering_head`.

Instead, when the VOPR restarts a replica, it suppresses storage faults during recovery if the replica entering `status=recovering_head` would possibly deadlock the cluster:

    // To improve VOPR utilization, try to prevent the replica from going into
    // `.recovering_head` state if the replica is needed to form a quorum.
    const fault = recoverable_count >= recoverable_count_min or replica.standby();
    simulator.replica_restart(replica.replica, fault);

Because corrupt WAL prepares can leak into "corrupt" (zeroed) WAL headers, when `replica_restart()` is invoked with faults disabled, it actually repairs the zeroed WAL headers:

    var header_prepare_view_mismatch: bool = false;
    if (!fault) {
        // The journal writes redundant headers of faulty ops as zeroes to ensure
        // that they remain faulty after a crash/recover. Since that fault cannot
        // be disabled by `storage.faulty`, we must manually repair it here to
        // ensure a cluster cannot become stuck in status=recovering_head.
        // See recover_slots() for more detail.
        const headers_offset = vsr.Zone.wal_headers.offset(0);
        const headers_size = vsr.Zone.wal_headers.size().?;
        const headers_bytes = replica_storage.memory[headers_offset..][0..headers_size];
        for (
            mem.bytesAsSlice(vsr.Header.Prepare, headers_bytes),
            replica_storage.wal_prepares(),
        ) |*wal_header, *wal_prepare| {
            if (wal_header.checksum == 0) {
                wal_header.* = wal_prepare.header;
            } else {
                if (wal_header.view != wal_prepare.header.view) {
                    header_prepare_view_mismatch = true;
                }
            }
        }
    }

This bug was hit by the VOPR with [misdirected write injection](https://github.com/tigerbeetle/tigerbeetle/pull/2463) as part of a liveness bug. I ran into it because the restart in question was due to an upgrade (`release_execute()`), and upgrades don't use the above redundant-header-repair.

# Fix

- Rather than ignoring `headers_redundant` during `Journal.header_sector()` when `faulty.bit(slot)`, maintain it properly so `Journal.header_sector()` can just use it as-is. That is, `headers_redundant` always contains "what we would write to the redundant headers, is we were to write to it right now".
- When `set_header_as_dirty` replaces a header, set the corresponding `headers_redundant` to an invalid header.
- During journal recovery, initialize `headers_redundant` with invalid headers corresponding to faulty prepares.

(We still set `headers_redundant` to the proper value in `Journal.write_prepare_header()`, just before generating the sector of redundant headers.)

Seed: ./zig/zig build -Drelease vopr_lite -- 13892479530912897922
Seed: ./zig/zig build -Drelease vopr_lite -- 3612761540040692234
Seed: ./zig/zig build -Drelease vopr_lite -- 18444273742674001259