vsr: reduce minimal backoff delay
The effective minimal delay for backoff is `rtt*2 + backoff_min`. This
is the lower-bound for a retry delay, and it should be proportional to
the time it takes to replicate a message across the ring of replicas.

For a fast network, transferring a 1MiB message between two replicas can
take as little as 500μs (https://github.com/sirupsen/napkin-math), or
~3ms to replicate across the entire ring.

The current minimum exponential backoff delay, 100ms, is way to
pessimistic for this scenario. Note that, for a slow network, the rtt
term will already account for slowness.