CFO: Split budget/timeout/upload intervals
# Bug

Previously, we only started new fuzzers during the first `budget_minutes` (10 minutes) of each CFO run.

At the end of that 10 minutes, we wait for up to another `hang_minutes` (30 minutes) to wait for any long running fuzzers to finish. But during this time we don't start any new fuzzers.

This that CFO cores are idle for up to 30 of every 40 minutes. This is also the rate at which the CFO pushes seeds to the devhub and refreshes its tasks.

# Fix

To fix:

- Remove `hang_minutes`.
- `budget_minutes` is now the total fuzz loop duration.
  - That is, it determines how often the CFO's code is refreshed.
  - (Approximately -- it doesn't count compile + upload times.)
  - This is now set to 40 minutes, to match our "real" budget from before. (I want it to be long enough to have a usefully-high `timeout_minutes`.)
- Upload seeds to the devhubdb (and in the follow-up commit, refresh our tasks) every `upload_minutes`. This is now set to 5 minutes.
- If a fuzzer runs for longer than `timeout_minutes`, it is killed and counts as a failure.

With this fix, CFO will use all cores all the time, and it will push updates every 5 minutes.
So when you add a fuzz label to a PR, the maximum-time-till-devhub-seeds is now 10 minutes instead of 80 minutes.