cfo: fix OOM handling
CFO already includes code to handle OOM:

    if (std.meta.eql(term, .{ .Signal = std.posix.SIG.KILL })) {
        // Something killed the fuzzer. This is likely OOM, so count this seed
        // neither as a success, nor as a failure.
        log.info("ignored SIGKILL for '{s}'", .{fuzzer.seed.command});

Sadly, this code is effectively a no-op. The problem here is that CFO
runs fuzzers via `zig build run`, an intermediate process. So, when oom
killer terminates the fuzzing process, the build_runner process notices
that and exits with non-zero status.

To fix this, arrange it so that we run the fuzzer process directly. To
that end, add `-Dprint-exe` option to build.zig to know _what_ to run
(we can't run the binaries from `./zig-out/bin`, because, eg, `vopr` and
`vopr -Dstate-machine=testing` alias).