VSR/Grid: Remove invalid zero-padding assert
In the grid, after we read a block, we:
1. verify that the block has a valid header, then
2. verify that the block has a valid body, then
3. verify that it is the block we were expecting to find, then
4. assert that the block's sector padding is zeroed.

The last step can fail due to disk corruption -- sector padding isn't covered by any checksum.

Despite injecting many storage faults the VOPR couldn't detect this (until the grandparent commit).
This is because when the VOPR injects a storage fault, it corrupts the whole sector -- if the padding is corrupt, then the header or body (or both) is also corrupt.

We missed this bug because we injected "too many" faults!

Seed: ./zig/zig build -Drelease vopr -- --lite 12441136989613272800