VSR/SuperBlock: Validate `checksum_padding`
Even though `checksum_padding` is reserved for future use, consider it part of the superblock checksum for the purpose of checksum validation.
Otherwise a bit error within the checksum will be ignored until we eventually assert that two copies of the same header are equal, and find that one of them has an invalid padding value.

Seed: ./zig/zig build -Drelease vopr -- --lite 505264887382715275