CFO: Refactor tasks, and fix another schedule weight bug
## Bug

Consider if you have a couple PR's being fuzzed (along with main), and then one of them gets merged.

When both PR's were being fuzzed, we split our fuzz time 50:50 between main and PR's.
So each PR got 25% of runtime. This is currently accomplished with task weight adjustments.

But when one of the PR's is merged, the fuzz time is still split 50:50 -- the remaining PR now has much higher (relative) weight than before.
That increase in weight means that the denominator of the `runtime / weight` is higher, so the remaining PR's virtual runtime is suddenly cut in half.
This would cause it to starve main until it catches up.

## Fix

Accumulate virtual runtime (`runtime_virtual`) as `+= 1000 / weight`, so that changes to the weight don't cause massive virtual runtime imbalances.

We still track regular `runtime` just because that is more meaningful for logging. But `runtime` isn't used for scheduling anymore.