Test/Storage: Don't misdirect WAL prepare if header is corrupt
Prevent a double-fault.

In this seed, the replica was recovering with its head op in case `@F` (valid prepare, invalid header, op_max).
The prepare in question was from an old view -- it was not overwritten due to a lost write.
Then that old header was truncated because it was not in our expected view-range.
This caused the replica to regress its op-head.

Seed: ./zig/zig build -Drelease vopr -- 2416633562060536000