inspect: start memory section
The "memory" section gives a napkin math estimate for the memory usage.
It is intentionally calculated manually from constants, rather than by
running inits with some kind of fake allocator, to double check that we
understand where the memory is going. That is, large discrepancies
between what `inspect constants` prints and what we get in actual usage
are possible, but it's not given that the bug is in inspect.

Start with giving the size of FreeSet.