node: don't crash when a client closed concurrent
Problem: if JS calls client.destroy() while there are outstanding
requests, the native part crashes with unreachable.

Cause: two bugs, actually!

_First_, `destroy` was using `napi_tsfn_abort` for cleaning-up the
trampoline used to invoke the JS callback. But before that,
client_deinint would actually enqueue several calls to that callback. We
only should be actually destroying it once all callbacks are fired, that
is, when node-side refcount reaches zero. That's exactly what release is
for!

_Second_, using `release` uncovered an off-by-one error in our reference
counting, with the function never released. That's because in the `init`
we end up with ref-count 2, we get one from

```
1, // Number of acquires/threads that will be calling this TSFN.
```

and a redundant one from explicit acquire, which is removed.

The relevant node docs are at

<https://nodejs.org/api/n-api.html#n_api_reference_counting_of_thread_safe_functions>

I wish I didn't have to write a manual test here, and that instead we
hade some sort of language-agnostic client test harness, some sort of
vörtex where you'll only have to write a client-specific driver, and
let the fuzzer find lifecycle bugs for you. But I'll leave that for
future work!

Closes: #2636