node: fix free of a wild pointer
In node client, we use packet.data_size for two purposes:

- originally, we store there a byte-size of the underlying buffer
- but then we store even&result count

So, like the best C code, we are reusing the same field to mean two
different things, and we switch between counts and sizes.

What could go wrong?

In this particular case, if an error occurs, then we don't switch the
meaning of the field and so, in `on_complete_js`, interpret len as
count, freeing the wrong slice as a result!

I've decided to apply a tactical fix here, but, really, I think this
whole thing needs to be re-done, its much to tricky.