trace: check that events' stack values don't collide
We manually assign a unique index to every event, so we must check that
the indices don't collide. To do this, we exhaustively enumerate all the
events and check that the indexes are indeed different.

To avoid writing repetitive nested loops, use the exhaustigen pattern
which allows the code for generating a "random" event to actually
enumerate all of the events.