vsr: assert that we are only sending valid prepare_oks
Yesterday, I panicked when reading this code as I interpret it in a
sense that we ack anything we have the _header_ for, ignoring the
prepares.

On the closer read, `send_prepare_ok` will withold prepare_ok if the
prepare is absent, but let's assert that at the call-site for clarity.