VSR: Explicit deprecated message types
## Bug

The `Command` enum lists all message types, with explicit values.
But if a replica receives a deprecated message type (e.g. via network replay), we would `switch` on the value, and panic.

Stack (from a replica on 0.16.28):

```
thread 24341 panic: switch on corrupt value
.../tigerbeetle/src/vsr/message_header.zig:144:17: 0x1304df3 in into_any (tigerbeetle)
.../tigerbeetle/src/vsr/message_header.zig:198:30: 0x18d9e98 in peer_type (tigerbeetle)
.../tigerbeetle/src/message_bus.zig:878:78: 0x18330e3 in set_and_verify_peer (tigerbeetle)
.../tigerbeetle/src/message_bus.zig:786:72: 0x17aeffc in parse_message (tigerbeetle)
.../tigerbeetle/src/message_bus.zig:732:48: 0x174e8e3 in parse_messages (tigerbeetle)
.../tigerbeetle/src/message_bus.zig:1007:42: 0x16edbab in on_recv (tigerbeetle)
.../tigerbeetle/src/io/linux.zig:1160:29: 0x16a6046 in wrapper (tigerbeetle)
.../tigerbeetle/src/io/linux.zig:686:40: 0x11f8ff5 in complete (tigerbeetle)
.../tigerbeetle/src/io/linux.zig:192:49: 0x11f775f in flush (tigerbeetle)
.../tigerbeetle/src/io/linux.zig:147:27: 0x1234796 in run_for_ns (tigerbeetle)
.../tigerbeetle/src/tigerbeetle/main.zig:510:38: 0x12362b4 in start (tigerbeetle)
.../tigerbeetle/src/tigerbeetle/main.zig:84:44: 0x12cfb3f in main (tigerbeetle)
.../tigerbeetle/zig/lib/std/start.zig:524:37: 0x11e6335 in posixCallMainAndExit (tigerbeetle)
.../tigerbeetle/zig/lib/std/start.zig:266:5: 0x11e5e51 in _start (tigerbeetle)
???:?:?: 0x4 in ??? (???)
Unwind information for `???:0x4` was not available, trace may be incomplete
```

## Fix

List deprecated message types in the `Command` enum, so that they are handled by switch statements.

Note that when we roll out a new message type, we need to make sure there is a transition release where the message type is received+ignored but not sent, otherwise we would be vulnerable to this bug in the other direction.