vsr: explicitly crash on unknown commands
I've spend a couple of hours in analysis-paralisys today (and like a day
cummulatievly before) over what to do with unknown messages.

I think it's best to preserve our current behavior of crashing, _but_ to
make it explicit, with a clear log message explaining _what_ unknown
command was received.

The motivation here is abundance of caution, and desier to avoid knight
capital kind of situation, where a replica _incorrectly_ ignores the
command.

One potential issue here is that we'll also crash on unknown commands
from clients, and we don't require clients to run with ECC ram. But the
thing is, if we recive an unknown command, we can't tell if it is from
the client or from the replica! So, yes, a bit flip in a client that
happens right before the client checksums can kill a replica, but that
should be fine, as replica will just restart.

We still don't have a good way to verify that this code works correctly,
so I manually patched our client to send an invalid command, and
verified that it does print:

```
2025-02-25 17:23:03.980Z error(message_bus): unknown vsr command, crashing for safety: '92'
thread 2892524 panic: unknown vsr command
```

Closes: #1850