vopr: fix two-orders-of-magnitude error in probabilities
The PRNG PR, #2798, introduced a bug, where a probability like `0.00002`
was converted to `ratio(2, 100_000)`. Such probabilites were being fed
into `chance_f64` function, which treated them as persentange. That is,
the probability was not `0.00002`, but rather `0.00002%`, a hundred
times more rare!

Add the zeroes back! replica_unpause_probability gets one extra zero,
which was genuinely typo'ed out in #2798.

With the current probabilities, the cluster exhausts AOF.