vopr: allow fast writes
In the current VOPR config, the minimum mean storage latency is 3 ticks,
which is 30 ms, which is just ridiculous. We generally expect almost all
IOPs to complete faster than in a single tick.

To model "faster than a tick" realistically, I change the simulator to
drain storage&network IO "within" each tick. As there could be multiple
immediately ready IOPs, the simulator randomizes the order in which
completions are invoked.