CFO: Retain checkouts + cache across iterations
Right now we force (each) CFO to recompile all fuzzers every iteration (so every ~hour) by deleting the working tree (which includes zig's binary cache).

This leads to a lot of extra compiling -- for the current PR's it is about 10 minutes for every 60 minutes of actual fuzzing.
Since the branches don't change that often, this is unnecessary work -- we'd rather spend those 10 minutes fuzzing instead.

(When we are compiling, we aren't starting fuzzers, and compilation is only single-threaded, so the CFO machine spends that ~10 minutes very underutilized.)

Instead let's retain the cache and just garbage-collect it selectively when tasks are removed.

Note that this change requires updating the supervisor script.