vopr: fix false positive
Fixes a vopr seed causing a  crash during upgrade. I belive this is a
pre-exising VOPR bug, which was uncovered by the tickless IO from #2823.

In production, release_execute is a `noreturn` function that immediately
re-execs into the newer version.

In VOPR, `release_execute` just sets the flag. VOPR acts on the flag
when it gets controll. Before #2823 storage.tick, and, after,
storage.step executed IOPs and next_tick_queue callbacks after each
other. So it was a possibility that an IOP triggers release_exectue, and
a next_tick steps into it, but it wasn't hit before.

To fix it, ensure that storage.step "yields" between IOPs and next
ticks. This is still perhaps not ideally, as we _should_ be yielding
after every _individual_ IOP, but lets see if this is enough to get the
thing fixed. I _think_ we should have a `next_tick` somewhere in commit
loop to make sure that the "yield" does happen.

Seed: ./zig/zig build vopr -Dvopr-state-machine=testing -- 18445183419870469729