vopr: properly fix upgrades during tickless IO
In #2834, I fixed vopr failure by introducing a "yield" into the step
function, between IOPs and next_tick, but that isn't really the right
way to do that.

Instead, we should set a flag to notify step that reset happened, so
that it can properly wind down.

This is how it used to work _before_ step, when `releae_execute_soon`
would simply `.reset` the storage. This worker for tick, which simply
procesesed events from the ready queue. But now that the step copies
ready events to a local variable, we need some more active reset
process.