vsr: fix crash in primary_repair_pipeline
The last step of the repair for prospective primary is pipeline repair
(loading prepares from WAL on disk into pipeline in memory). It is
tricky because, when WAL read finishes, the replica might no longer be a
primary!

So `repair_pipeline_read_callback` starts with a bunch of checks to
ensure that we still should be repairing pipeline. It misses one case,
when there's a header break betwene commit_max and op.

Seed: ./zig/zig build  vopr -Drelease -- 16932172197309204889