fuzz: remove global allocator
There's no reason for us to use a global allocator during fuzzing. Let's
align this more with TigerStyle and pass the allocator explicitly: it's
not really any more code, but it makes it clear where the allocation can
happen, and where we are violating "allocate only during startup" rule,
which is not _critical_ for fuzzing, but helps non-the-less!