vopr: re-do tickless VOPR
This is a second take at #2823. The idea is the same: with many IOPs
ready at any given tick, we want to randomize the order in which IOPs
are executed.

The implementation is different --- rather than collecting a list of
ready IOPS and shuflling it, I add a helper function to pick a random
IOP directly off the queue. This keeps the queue as the real source of
truth about pending operations.

The "source of truth property" is important:

* For upgrades, where we clear the queue and expect that that removes
  all operations.
* For packet simulator, where we size message pool accoring to the size
  of the queue.