[VSR] MultiBatch encoder and decoder Implements the message encoder and decoder supporting the multi-batch format.
Although TigerBeetle supports batches as first-class primitives, sometimes applications need to submit independent units of work concurrently through the same client, creating the opportunity to pack them together into the same physical message to be executed within the same `op`.

MultiBatching solves the problem of how the client and the StateMachine should encode and decode these batches inside the message body by adding a thin trailer at the end of the message.
The trailer contains the number of batches encoded and the number of elements in each batch, and has a variable size, always a multiple of the operation's element size.

See the comments on `multi_batch.zig` for more details.