[StateMachine] MultiBatch support The state machine supports MultiBatch-encoded messages and replies, as well as legacy non-encoded messages for backwards compatibility.
Decoding and encoding are handled by the state machine (which is aware of multi-batching) and can optimize operations such as `prefetch` in a single pass for the entire payload (regardless of the number of batches encoded).
All operations support multi-batch encoding, even queries that require a single element per batch may have multiple filters encoded in the message body (bounded by `filter.limit`), handling operations where the max number of results does not depend on the number of events.

- Unit tests for both encoded and legacy messages.
- Unit tests for validate batches inputs.
- Benchmark and Repl