vopr: refactor packet simulator
PacketSimulator is generic over packet. Previously, polymorphism was
expressed via:

* calling clone/deint functions on the packet, duckly-typed
* passing-in a callback for each submit

In this refactor, we invect a VTable for dropping/cloning/delivering a
packet at init. The VTable is runtim, but this doesn't really matter,
could have been comptime.

The _immediate_ benefit here, and the original motivation, was that now
the Network can use a raw message as a packet, and not a
message,*Network pair. In turn, this allows a killer feature: packet
simulator now prints a packet when dropping it, making debugging easier!
(we don't print packet on delivery, as we expect the replica to print it
anyway).

But I actually eyed this code with suspicion for a long time --- I just
don't _like_ how we do polymorphism here, a vtable seems a just a nicer,
more well-rounded approach.