vsr: call hook into vopr while sending message to client
In 5de082c227b4ebb72aa11945a132004cb0bc5e26, we added new logic to
track replies based on when they are externalized from replica to
client. Turns out this new logic was not being exercised at all,
since we only call the hook into the VOPR when messages are sent to
other replicas, not when they are sent to the client!

Fix it by calling the hook when messages are sent to the client,
this would lead to reply messages getting appropriately tracked.

Seed: ./zig/zig build -Drelease vopr -Dvopr-state-machine=testing -- 10711934386281284252