clients: add tb_client_init_parameters and implement for python
This would have caught #2652 - it adds a simple round trip method,
tb_client_init_parameters, that allows retrieving the cluster_id and addresses
passed in when creating a cluster.

They are stored on the client context, so it works with both the real and echo
clients. There's a little bit of overhead doing it this way, but I think it's
the most straighforward.

Addresses in particular is duplicated as a string - IPv6 addresses (or even
v4) might canonicalize and thus differ from the value passed in if the
net.Address were formatted instead.

It also adds a test for Python - which should be extended to all languages.