[StateMachine] `get_events` operation.
Defines the event filter and result for the `get_events` operation.
The filter supports basic pagination for CDC: `timestamp + limit`,
always in ascending order.

The result format is based on `AccountEvents`, but includes the full contents
of both `transfers` and `accounts` objects, along with the balance *as-of* timestamp.

Expiry events return the original transfer, so `event.timestamp != event.transfer_timestamp` in this case.