git-review: add split subcommand
The idea is that you can edit the code in-place during reviewing, than
add everything, comments and code changes, into the review commit, and
then call `git review split` to split the commit in two, one with code
changes and one with just the review comments.

The main benefit here is that this ordres code changes _before_ the
review, so any further history rewriting becomes easier.

In other words, in a typical workflow wheer you make code-changes as a
commint atop of the review commit, rebasing this commit to be before the
review would result in nasty merge conflicts, as you are likely to
modify adjacent lines. The `git-review split` does the right thing
there.