cfo: add arbitrary debug output to seeds
To debug false positives, we want to get at least the exit code. I don't
want to add `exit_code` field though, as I intentionally don't want this
to be structured, and I don't want to further change the format. So, we
use a general purpose `debug` string instead.

I'd also love to catch last 1KiB of stderr here ideally, but, as far as
I can tell, there isn't an easy way to do this, which is a shame --- a
ring buffer device sounds neat and very UNIX-y. So that's left for
future work.