git-review: add command to send review comments as a email
`git-review email --recepient somebody@example.com` will parse the
current state of review and send a plain-text email to the specified
address. Currently, this is mostly just scaffolding. I confirmd that it
can, in fact, send an email, but there's a bunch of things yet to do to
make it fully useful:

- setup a review bot email account (I've been using my personal account
  for testing)
- actually call `git review send` in GHA
- add a gitdb to skip already emailed comments
- add nicer formatting for the actual mail

But I want to move one step at a time here, and the first step is
crafting of the relevant CURL command.