io: let the caller drive run_for_ns inner loop
Previously, the main loop looked like this:

    while(true) {
       tick()
       io.run_for_ns(10ms);
    }

After this commit, it is instead:

    while(true) {
       tick()
       io.run_for_ns_setup(10ms);
       while (try io.run_for_ns(10ms)) {
       }
    }

Notably, inside while loop the caller knows that there aren't any
pending ready callbacks, so this is a useful place to "poll" for things.

This setup is a pre-requisite for making message bus poll-based, to
follow not react to stimuli principle.

An alternative design would be to keep the loop inside `run_for_ns`,
_but_ allow configuring an io instance with some sort of `idle`
callback. I intentionally avoid that to simplify control flow.

I'll modify call-sites in the next commit, when message bus actually
gains poll-based interface