bus: pull-based message bus
Before this comment, message bus invoked an on_message callback every
time a recv syscall returned enough bytes to form a valid message. Now,
messages need to be explicitly pulled out of the bus via receive_message
method.

 ## Background

Pulling events is in general a good pattern for asynchronous systems.
Imagine a generic evented system which works like:

    for message in incoming_messages:
        process(message)

Note that because `process` takes time, by the time it finishes and we
are ready to pull the next message from the pile, there might be _many_
messages in the pile already! And usually there are _some_ things to be
gained from exploiting the pile as a whole, rather than looking at it as
just a sequence of messages:

* You can prioritize, and select the more important messages out of the
  pile, even if they arrived later.
* You can coalesce and remove "duplicate" messages (e.g., taking only
  the latest state update).
* You can batch message processing. For example, processing all updates
  in one go and sending only a single change notification.

These points don't significantly apply to TigerBeetle, but there's one
issue that does apply, backpressure!

  ## Precarious Prepare Pileup

In the today's push-based setup, Replica's `on_message` is called
immediately upon receiving the message, and the replica might deal with
the message immediately. This is not always possible, the replica might
lack the required resources.

One very specific instance of this that we observed in real-world
deployments is message bus receiving a whole bunch of tiny prepares in
one go, lets say, 20 of them. When that happens, the bus calls
on_message 20 times in a row. But because we only have 8 IOPs in the
journal, most of the prepares are dropped on the floor, which creates
then causes extra repair and prepare traffic.

The behavior we want here is for the replica to _only_ pull a prepare
message out of the message bus when it has a spare IOP, and that's
exactly what's achieved here.

  ## Implementation

The core of the change is that MessageBus loses the `on_message`
callback, and instead gains a

  fn receive_message(bus: *Bus) ?*Message

instead. The replcia can now poll the bus when it is ready to process a
message. The fully general solution here would expose something like

   const MessageIterator = struct {
       fn peek(it: *const MessageIterator) ?*const Header,
       fn next_take(it: *MessageIterator) *Message,
       fn next_skip(it: *MessageIterator) *const Message,
   }

to allow the replica to inspect all available messages and receive them
selectively, but:

- I don't think we need to go _that_ far
- and, even if we need, it'll be a relatively uninteresting diff on top
  of the present commit.

So instead there's just `peek_message`, `receive_message`, which only
allow stalling the processing, rather than being selective. And we only
stall if we don't have a write IOP available, so the stalls should be
short, and I wouldn't worry about us delaying pings too much.

The interesting question is _when_ should the replica call
`bus.receive_message()`? How do we notify about readiness? The proposed
answer is "all the time". On every iteration of the event loop, before
we go into the kernel to sleep, we invoke

    fn idle(replica: *Replica) void

callback which then polls the bus.

The nice thing about this solution is that we avoid function pointes or
weird stacks: `replcia.idle()` is a _direct_ call, and we do it right
from main, there's nothing above us on the stack.

The not-so-nice thing is that this _seems_ inefficient. The previous
approach, with on_message callback, notified us precisely! We can
restore that by keeping the on_message callback, but then letting the
replica poll for the message from that callback. This would be an
incomplete solution! If the replica decides to postpone receiving a
prepare, it still needs to poll the bus later, when it is ready to
process the message. So the full solution here would be to pull from the
bus in the on_message callback, and _also_ in the `write_prepare` /
`write_block` callback. But that would create ugly backtraces, where
there's unrelated stuff above in the stack. To fix that, one option
would be just to set a `need_to_pull_message_bus` _flag_ in the
`write_prepare_callback`. And then, when we have nothing else on the
call-stack, we can check the flag and poll the bus. But then this
becomes equivalent to the idle solution!

This line of thinking convinced myself that the loop with `.idle` in
main is the right approach.

What I _don't_ like about this solution is that it make the VOPR event
loop even more complex, but, given that the production loop looks fine,
I am ok with that.